Remove queue integration tests
Change-type: patch
