+++ test/ci-tasks.yml
@@ -143,11 +143,6 @@ tasks:
         value: './test/integration/sync/outreach-translate.spec.js'
     dependencies:
       - 'typeform-translate'
-  - name: 'integration-queue'
-    description: 'Queue Integration Tests'
-    command: 'make test-integration-queue'
-    cwd: '/usr/src/jellyfish'
-    required: true
   - name: 'integration-server'
     description: 'Server Integration Tests'
     command: 'make test-integration-server'
@@ -156,8 +151,6 @@ tasks:
     environment:
       - name: 'SERVER_HOST'
         value: 'http://localhost'
-    dependencies:
-      - 'integration-queue'
   - name: 'integration-worker'
     description: 'Worker Integration Tests'
     command: 'make test-integration-worker'
