+++ test/integration/worker/actions/complete-first-time-login.spec.js
@@ -7,7 +7,6 @@
 const ava = require('ava')
 const nock = require('nock')
 const helpers = require('../helpers')
-const actionLibrary = require('@balena/jellyfish-action-library')
 const uuid = require('@balena/jellyfish-uuid')
 const environment = require('@balena/jellyfish-environment')
 
@@ -97,7 +96,7 @@ const createUser = async (test, withPassword) => {
 }
 
 ava.before(async (test) => {
-	await helpers.worker.before(test, actionLibrary)
+	await helpers.worker.before(test)
 	const {
 		session,
 		jellyfish,
