+++ test/integration/worker/actions/send-first-time-login-link.spec.js
@@ -8,7 +8,6 @@ const ava = require('ava')
 const nock = require('nock')
 const helpers = require('../helpers')
 const uuid = require('@balena/jellyfish-uuid')
-const actionLibrary = require('@balena/jellyfish-action-library')
 const environment = require('@balena/jellyfish-environment')
 
 const MAIL_OPTIONS = environment.mail.options
@@ -52,7 +51,7 @@ const createOrgLinkAction = async ({
 }
 
 ava.before(async (test) => {
-	await helpers.worker.before(test, actionLibrary)
+	await helpers.worker.before(test)
 	const {
 		jellyfish,
 		context,
